requirements.txt
pandas>=0.21
requests>=2.3.0
requests-cache>=0.5.0
python-decouple